+++ typescript/packages/plugins/orca/src/index.ts
@@ -1,2 +1,1 @@
 export * from "./orca.plugin";
-export * from "./parameters";
