+++ typescript/packages/wallets/solana/src/SolanaKeypairWalletClient.ts
@@ -1,19 +1,11 @@
-import {
-    type Keypair,
-    TransactionMessage,
-    VersionedTransaction,
-} from "@solana/web3.js";
+import { type Keypair, TransactionMessage, VersionedTransaction } from "@solana/web3.js";
 import nacl from "tweetnacl";
-import {
-    type SolanWalletClientCtorParams,
-    SolanaWalletClient,
-} from "./SolanaWalletClient";
+import { type SolanWalletClientCtorParams, SolanaWalletClient } from "./SolanaWalletClient";
 import type { SolanaTransaction } from "./types";
 
-export type SolanaKeypairWalletClientCtorParams =
-    SolanWalletClientCtorParams & {
-        keypair: Keypair;
-    };
+export type SolanaKeypairWalletClientCtorParams = SolanWalletClientCtorParams & {
+    keypair: Keypair;
+};
 
 export class SolanaKeypairWalletClient extends SolanaWalletClient {
     #keypair: Keypair;
@@ -30,31 +22,20 @@ export class SolanaKeypairWalletClient extends SolanaWalletClient {
 
     async signMessage(message: string) {
         const messageBytes = Buffer.from(message);
-        const signature = nacl.sign.detached(
-            messageBytes,
-            this.#keypair.secretKey
-        );
+        const signature = nacl.sign.detached(messageBytes, this.#keypair.secretKey);
         return {
             signature: Buffer.from(signature).toString("hex"),
         };
     }
 
-    async sendTransaction({
-        instructions,
-        addressLookupTableAddresses = [],
-        accountsToSign = [],
-    }: SolanaTransaction) {
+    async sendTransaction({ instructions, addressLookupTableAddresses = [], accountsToSign = [] }: SolanaTransaction) {
         const latestBlockhash = await this.connection.getLatestBlockhash();
 
         const message = new TransactionMessage({
             payerKey: this.#keypair.publicKey,
             recentBlockhash: latestBlockhash.blockhash,
             instructions,
-        }).compileToV0Message(
-            await this.getAddressLookupTableAccounts(
-                addressLookupTableAddresses
-            )
-        );
+        }).compileToV0Message(await this.getAddressLookupTableAccounts(addressLookupTableAddresses));
         const transaction = new VersionedTransaction(message);
 
         transaction.sign([this.#keypair, ...accountsToSign]);
@@ -70,7 +51,7 @@ export class SolanaKeypairWalletClient extends SolanaWalletClient {
                 lastValidBlockHeight: latestBlockhash.lastValidBlockHeight,
                 signature: hash,
             },
-            "confirmed"
+            "confirmed",
         );
 
         return {
@@ -79,5 +60,4 @@ export class SolanaKeypairWalletClient extends SolanaWalletClient {
     }
 }
 
-export const solana = (params: SolanaKeypairWalletClientCtorParams) =>
-    new SolanaKeypairWalletClient(params);
+export const solana = (params: SolanaKeypairWalletClientCtorParams) => new SolanaKeypairWalletClient(params);
